Fix Metal type mapping: Coex int/float are 64-bit
Coex int is 64-bit (i64), so must map to Metal 'long' not 'int'.
Coex float is 64-bit (f64), so must map to Metal 'double' not 'float'.

This fixes GPU kernel execution where only some elements were
processed correctly - the 32-bit/64-bit mismatch caused the kernel
to read/write wrong memory offsets.

+++ codegen/formula/metal.py
@@ -18,11 +18,12 @@ class MetalBackend(FormulaBackend):
     """
 
     # Coex type to Metal type mapping
+    # IMPORTANT: Coex int is 64-bit, so maps to Metal 'long' (int64_t)
     TYPE_MAP = {
-        'int': 'int',
+        'int': 'long',         # Coex int is 64-bit
         'int64': 'long',
-        'float': 'float',
-        'float64': 'double',  # Note: check device capability
+        'float': 'double',     # Coex float is 64-bit
+        'float64': 'double',
         'bool': 'bool',
         'byte': 'uint8_t',
     }
